refactor: split multi-exporter files
