+++ addons/io_scene_gltf2_msfs/io/msfs_multi_export.py
@@ -21,53 +21,6 @@ import uuid
 import xml.dom.minidom
 import xml.etree.ElementTree as etree
 
-# Property Groups
-class MultiExporterObjectLOD(bpy.types.PropertyGroup):
-    object: bpy.props.PointerProperty(name="", type=bpy.types.Object)
-    enabled: bpy.props.BoolProperty(name="", default=False)
-    lod_value: bpy.props.IntProperty(name="", default=0, min=0, max=1000)
-    flatten_on_export: bpy.props.BoolProperty(name="", default=False)
-    keep_instances: bpy.props.BoolProperty(name="", default=False)
-    file_name: bpy.props.StringProperty(name="", default="")
-
-
-class MultiExporterObjectGroup(bpy.types.PropertyGroup):
-    group_name: bpy.props.StringProperty(name="", default="")
-    expanded: bpy.props.BoolProperty(name="", default=True)
-    lods: bpy.props.CollectionProperty(type=MultiExporterObjectLOD)
-    folder_name: bpy.props.StringProperty(name="", default="", subtype="DIR_PATH")
-    generate_xml: bpy.props.BoolProperty(name="", default=True)
-
-
-class MultiExporterPresetLayer(bpy.types.PropertyGroup):
-    collection: bpy.props.PointerProperty(name="", type=bpy.types.Collection)
-    enabled: bpy.props.BoolProperty(name="", default=False)
-    expanded: bpy.props.BoolProperty(name="", default=True)
-
-
-class MultiExporterPreset(bpy.types.PropertyGroup):
-    def update_file_path(self, context):
-        # Make sure file_path always ends in .gltf
-        if os.path.basename(self.file_path):
-            file_path = bpy.path.ensure_ext(
-                os.path.splitext(self.file_path)[0],
-                ".gltf",
-            )
-            if self.file_path != file_path:
-                self.file_path = file_path
-        elif not os.path.isfile(self.file_path):
-            file_path = os.path.join(self.file_path, self.name + ".gltf")
-            if self.file_path != file_path:
-                self.file_path = file_path
-
-    name: bpy.props.StringProperty(name="", default="")
-    file_path: bpy.props.StringProperty(
-        name="", default="", subtype="FILE_PATH", update=update_file_path
-    )
-    enabled: bpy.props.BoolProperty(name="", default=False)
-    expanded: bpy.props.BoolProperty(name="", default=True)
-    layers: bpy.props.CollectionProperty(type=MultiExporterPresetLayer)
-
 
 # Scene Properties
 class MSFSMultiExporterProperties:
@@ -150,7 +103,7 @@ class MSFS_OT_MultiExportGLTF2(bpy.types.Operator):
 
                         bpy.ops.export_scene.gltf(
                             export_format="GLTF_SEPARATE",
-                            export_selected=True,
+                            use_selection=True,
                             filepath=os.path.join(
                                 object_group.folder_name,
                                 os.path.splitext(lod.file_name)[0],
@@ -173,85 +126,11 @@ class MSFS_OT_MultiExportGLTF2(bpy.types.Operator):
 
                     bpy.ops.export_scene.gltf(
                         export_format="GLTF_SEPARATE",
-                        export_selected=True,
+                        use_selection=True,
                         filepath=os.path.join(preset.file_path),
                     )
 
         return {"FINISHED"}
-
-
-class MSFS_OT_ReloadObjectGroups(bpy.types.Operator):
-    bl_idname = "msfs.reload_object_groups"
-    bl_label = "Reload object groups"
-
-    def get_group_from_object_name(self, object_name):
-        matches = re.findall(
-            "(?i)x[0-9]_|_lod[0-9]+", object_name
-        )  # If an object starts with xN_ or ends with _LODN, treat as an LOD
-        if matches:
-            # Get base object group name from object
-            for match in matches:
-                filtered_string = object_name.replace(match, "")
-            return filtered_string
-        else:
-            # If prefix or suffix isn't found, use the object name as the group
-            return object_name
-
-    def execute(self, context):
-        object_groups = bpy.context.scene.msfs_multi_exporter_object_groups
-
-        # Remove deleted objects and empty object groups
-        for i, object_group in enumerate(object_groups):
-            for j, lod in enumerate(object_group.lods):
-                if not lod.object.name in bpy.context.scene.objects:
-                    object_groups[i].lods.remove(j)
-
-                # Make sure object still matches group name
-                if (
-                    not self.get_group_from_object_name(lod.object.name)
-                    == object_group.group_name
-                ):
-                    object_groups[i].lods.remove(j)
-
-            if len(object_group.lods) == 0:
-                object_groups.remove(i)
-
-        # Search all objects in scene to find object groups
-        found_object_groups = {}
-        for obj in bpy.context.scene.objects:
-            if obj.parent is None:  # Only search "root" objects
-                group_name = self.get_group_from_object_name(obj.name)
-
-                # Set object group or append
-                if group_name in found_object_groups.keys():
-                    found_object_groups[group_name].append(obj)
-                else:
-                    found_object_groups[group_name] = [obj]
-
-        # Create object groups and LODs
-        for _, (object_group_name, objects) in enumerate(found_object_groups.items()):
-            # Check if object group already exists, and if it doesn't, create one
-            if not object_group_name in [
-                object_group.group_name for object_group in object_groups
-            ]:
-                object_group = object_groups.add()
-                object_group.group_name = object_group_name
-            else:
-                for object_group in object_groups:
-                    if object_group.group_name == object_group_name:
-                        break
-
-            # Set all LODs in object group
-            for obj in objects:
-                # If the object is at the root level (no parent)
-                if obj.parent is None:
-                    if not obj in [lod.object for lod in object_group.lods]:
-                        lod = object_group.lods.add()
-                        lod.object = obj
-                        lod.file_name = obj.name
-
-        return {"FINISHED"}
-
 
 class MSFS_OT_ChangeTab(bpy.types.Operator):
     bl_idname = "msfs.multi_export_change_tab"
@@ -263,101 +142,6 @@ class MSFS_OT_ChangeTab(bpy.types.Operator):
         context.scene.msfs_multi_exporter_current_tab = self.current_tab
         return {"FINISHED"}
 
-
-class MSFS_OT_AddPreset(bpy.types.Operator):
-    bl_idname = "msfs.multi_export_add_preset"
-    bl_label = "Add preset"
-
-    def execute(self, context):
-        presets = bpy.context.scene.msfs_multi_exporter_presets
-        preset = presets.add()
-        preset.name = f"Preset {len(presets)}"
-        preset.file_path = preset.name + ".gltf"
-
-        return {"FINISHED"}
-
-
-class MSFS_OT_RemovePreset(bpy.types.Operator):
-    bl_idname = "msfs.multi_export_remove_preset"
-    bl_label = "Remove preset"
-
-    preset_index: bpy.props.IntProperty()
-
-    def execute(self, context):
-        presets = bpy.context.scene.msfs_multi_exporter_presets
-        presets.remove(self.preset_index)
-
-        return {"FINISHED"}
-
-
-class MSFS_OT_EditLayers(bpy.types.Operator):
-    bl_idname = "msfs.multi_export_edit_layers"
-    bl_label = "Edit layers"
-
-    preset_index: bpy.props.IntProperty()
-
-    collection_tree = {}
-
-    def execute(self, context):
-        return {"FINISHED"}
-
-    def getChildren(self, collection, children):
-        children[collection] = {}
-        for child in collection.children:
-            children[collection] = self.getChildren(child, children[collection])
-        return children
-
-    def invoke(self, context, event):
-        preset = bpy.context.scene.msfs_multi_exporter_presets[self.preset_index]
-
-        for i, layer in enumerate(preset.layers):
-            if not layer.collection in list(bpy.data.collections):
-                preset.layers.remove(i)
-
-        for collection in bpy.data.collections:
-            if not collection in [layer.collection for layer in preset.layers]:
-                layer = preset.layers.add()
-                layer.collection = collection
-
-        # Because it isn't really possible to define children in the layers, we have to generate a "tree" of collections with their children, and use that when rendering.
-        self.collection_tree = self.getChildren(bpy.context.scene.collection, {})
-
-        wm = context.window_manager
-        return wm.invoke_props_dialog(self)
-
-    def draw(self, context):
-        layout = self.layout
-
-        preset = bpy.context.scene.msfs_multi_exporter_presets[self.preset_index]
-
-        # Loop through our collection tree and draw layers with respect to children
-        def drawTree(layout_item, tree):
-            for i, (collection, children) in enumerate(tree.items()):
-                for layer in preset.layers:
-                    if layer.collection == collection:
-                        box = layout_item.box()
-                        row = box.row()
-                        if layer.collection.children:
-                            row.prop(
-                                layer,
-                                "expanded",
-                                text=layer.collection.name,
-                                icon="DOWNARROW_HLT"
-                                if layer.expanded
-                                else "RIGHTARROW",
-                                icon_only=True,
-                                emboss=False,
-                            )
-                            row.prop(layer, "enabled", text="Enabled")
-                            if layer.expanded:
-                                drawTree(box, children)
-                        else:
-                            row.label(text=layer.collection.name)
-                            row.prop(layer, "enabled", text="Enabled")
-
-                        break
-
-        drawTree(layout, self.collection_tree[bpy.context.scene.collection])
 
 
 # Panels
@@ -391,142 +175,7 @@ class MSFS_PT_MultiExporter(bpy.types.Panel):
         ).current_tab = "PRESETS"
 
 
-class MSFS_PT_MultiExporterObjectsView(bpy.types.Panel):
-    bl_label = ""
-    bl_parent_id = "MSFS_PT_MultiExporter"
-    bl_space_type = "VIEW_3D"
-    bl_region_type = "UI"
-    bl_category = "Multi-Export glTF 2.0"
-    bl_options = {"HIDE_HEADER"}
-
-    @classmethod
-    def poll(cls, context):
-        return context.scene.msfs_multi_exporter_current_tab == "OBJECTS"
-
-    def draw(self, context):
-        layout = self.layout
-
-        layout.operator(MSFS_OT_ReloadObjectGroups.bl_idname, text="Reload LODs")
-        layout.prop(context.scene, "multi_exporter_show_hidden_objects")
-
-        object_groups = context.scene.msfs_multi_exporter_object_groups
-
-        total_lods = 0
-        for object_group in object_groups:
-            for lod in object_group.lods:
-                if (
-                    not context.scene.multi_exporter_show_hidden_objects
-                    and lod.object.hide_get()
-                ):
-                    continue
-
-                total_lods += 1
-
-        if total_lods == 0:
-            box = layout.box()
-            box.label(text="No LODs found in scene")
-        else:
-            for object_group in object_groups:
-                row = layout.row()
-                if (
-                    len(object_group.lods) == 1
-                ):  # If we only have one LOD in the group, and it is hidden, then don't render the group
-                    if (
-                        not context.scene.multi_exporter_show_hidden_objects
-                        and object_group.lods[0].object.hide_get()
-                    ):
-                        continue
-
-                if len(object_group.lods) > 0:
-                    box = row.box()
-                    box.prop(
-                        object_group,
-                        "expanded",
-                        text=object_group.group_name,
-                        icon="DOWNARROW_HLT" if object_group.expanded else "RIGHTARROW",
-                        icon_only=True,
-                        emboss=False,
-                    )
-                    if object_group.expanded:
-                        box.prop(object_group, "generate_xml", text="Generate XML")
-                        box.prop(object_group, "folder_name", text="Folder")
-
-                        col = box.column()
-                        for lod in object_group.lods:
-                            if (
-                                not context.scene.multi_exporter_show_hidden_objects
-                                and lod.object.hide_get()
-                            ):
-                                continue
-
-                            row = col.row()
-                            row.prop(lod, "enabled", text=lod.object.name)
-                            subrow = row.column()
-                            subrow.prop(lod, "lod_value", text="LOD Value")
-                            # subrow.prop(lod, "flatten_on_export", text="Flatten on Export") # Disable these two options for now as there's not a great way to implement them
-                            # subrow.prop(lod, "keep_instances", text="Keep Instances")
-                            subrow.prop(lod, "file_name", text="File Name")
-
-        row = layout.row(align=True)
-        row.operator(MSFS_OT_MultiExportGLTF2.bl_idname, text="Export")
-
-
-class MSFS_PT_MultiExporterPresetsView(bpy.types.Panel):
-    bl_label = ""
-    bl_parent_id = "MSFS_PT_MultiExporter"
-    bl_space_type = "VIEW_3D"
-    bl_region_type = "UI"
-    bl_category = "Multi-Export glTF 2.0"
-    bl_options = {"HIDE_HEADER"}
-
-    @classmethod
-    def poll(cls, context):
-        return context.scene.msfs_multi_exporter_current_tab == "PRESETS"
-
-    def draw(self, context):
-        layout = self.layout
-
-        layout.operator(MSFS_OT_AddPreset.bl_idname, text="Add Preset")
-
-        presets = bpy.context.scene.msfs_multi_exporter_presets
-        for i, preset in enumerate(presets):
-            row = layout.row()
-            box = row.box()
-            box.prop(
-                preset,
-                "expanded",
-                text=preset.name,
-                icon="DOWNARROW_HLT" if preset.expanded else "RIGHTARROW",
-                icon_only=True,
-                emboss=False,
-            )
-
-            if preset.expanded:
-                box.prop(preset, "enabled", text="Enabled")
-                box.prop(preset, "name", text="Name")
-                box.prop(preset, "file_path", text="Export Path")
-
-                box.operator(
-                    MSFS_OT_EditLayers.bl_idname, text="Edit Layers"
-                ).preset_index = i
-
-                box.operator(
-                    MSFS_OT_RemovePreset.bl_idname, text="Remove"
-                ).preset_index = i
-
-        row = layout.row()
-        row.operator(MSFS_OT_MultiExportGLTF2.bl_idname, text="Export")
-
-
 def register():
-    bpy.types.Scene.msfs_multi_exporter_object_groups = bpy.props.CollectionProperty(
-        type=MultiExporterObjectGroup
-    )
-    bpy.types.Scene.msfs_multi_exporter_presets = bpy.props.CollectionProperty(
-        type=MultiExporterPreset
-    )
-
-    # Settings for multi exporter
     bpy.types.Scene.multi_exporter_show_hidden_objects = bpy.props.BoolProperty(
         name="Show hidden objects", default=True
     )
